feat: moved serialization code to backend separate from entities
