+++ common/entities/auth/group-membership.entity.ts
@@ -37,26 +37,4 @@ export default class GroupMembership extends BaseEntity {
      */
     @Column({ default: false })
     canEditGroup!: boolean;
-
-    get groupMembershipDto(): GroupMembershipDto {
-        return groupMembershipEntityToDto(this);
-    }
 }
-
-export const groupMembershipEntityToDto = (
-    groupMembership: GroupMembership,
-): GroupMembershipDto => {
-    if (groupMembership.user === undefined) {
-        throw new Error('Member can never be undefined');
-    }
-
-    return {
-        uuid: groupMembership.uuid,
-        canEditGroup: groupMembership.canEditGroup,
-        accessGroup: null, // we don't want to return the access group
-        createdAt: groupMembership.createdAt,
-        updatedAt: groupMembership.updatedAt,
-        expirationDate: groupMembership.expirationDate ?? null,
-        user: groupMembership.user.userDto,
-    };
-};
