+++ common/entities/auth/project-access.entity.ts
@@ -22,29 +22,4 @@ export default class ProjectAccess extends BaseEntity {
         nullable: false,
     })
     project?: Project;
-
-    get projectAccessDto() {
-        return projectAccessEntityToDto(this);
-    }
 }
-
-export const projectAccessEntityToDto = (projectAccess: ProjectAccess) => {
-    if (projectAccess.accessGroup === undefined) {
-        throw new Error('Access group not found');
-    }
-
-    if (projectAccess.accessGroup.memberships === undefined) {
-        throw new Error('Access group has no memberships');
-    }
-
-    return {
-        createdAt: projectAccess.createdAt,
-        hidden: false,
-        memberCount: projectAccess.accessGroup.memberships.length,
-        type: projectAccess.accessGroup.type,
-        updatedAt: projectAccess.updatedAt,
-        name: projectAccess.accessGroup.name,
-        rights: projectAccess.rights,
-        uuid: projectAccess.accessGroup.uuid,
-    };
-};
