+++ common/entities/file/file.entity.ts
@@ -74,53 +74,4 @@ export default class FileEntity extends BaseEntity {
 
     @Column({ nullable: true })
     origin?: FileOrigin;
-
-    get fileDto(): FileDto {
-        return fileEntityToDto(this);
-    }
-
-    get fileWithTopicDto(): FileWithTopicDto {
-        return fileEntitiyToDtoWithTopic(this);
-    }
 }
-
-export const fileEntityToDto = (file: FileEntity): FileDto => {
-    if (!file.creator) {
-        throw new Error('File creator is not set');
-    }
-
-    if (!file.mission) {
-        throw new Error('File mission is not set');
-    }
-
-    return {
-        uuid: file.uuid,
-        state: file.state,
-        createdAt: file.createdAt,
-        updatedAt: file.updatedAt,
-        date: file.date,
-        filename: file.filename,
-        type: file.type,
-        size: file.size ?? 0,
-        hash: file.hash ?? '',
-        creator: file.creator.userDto,
-        mission: file.mission.missionDto,
-        categories:
-            file.categories?.map((c) => ({
-                uuid: c.uuid,
-                name: c.name,
-            })) ?? [],
-    };
-};
-
-export const fileEntitiyToDtoWithTopic = (
-    file: FileEntity,
-): FileWithTopicDto => {
-    if (!file.topics) {
-        throw new Error('File topics are not set');
-    }
-    return {
-        ...fileEntityToDto(file),
-        topics: file.topics.map((topic) => topic.topicDto),
-    };
-};
