+++ common/entities/mission/mission.entity.ts
@@ -8,13 +8,6 @@ import Apikey from '../auth/apikey.entity';
 import Tag from '../tag/tag.entity';
 import MissionAccess from '../auth/mission-access.entity';
 import Action from '../action/action.entity';
-import {
-    FlatMissionDto,
-    MinimumMissionDto,
-    MissionDto,
-    MissionWithCreatorDto,
-    MissionWithFilesDto,
-} from '../../api/types/mission/mission.dto';
 
 @Unique('unique_mission_name_per_project', ['name', 'project'])
 @Entity()
@@ -47,91 +40,4 @@ export default class Mission extends BaseEntity {
 
     @OneToMany(() => Tag, (tag) => tag.mission)
     tags?: Tag[];
-
-    get missionDto(): MissionDto {
-        return missionEntityToDto(this);
-        
-    }
-
-    get missionWithCreatorDto(): MissionWithCreatorDto {
-        return missionEntityToDtoWithCreator(this);
-    }
-
-    get flatMissionDto(): FlatMissionDto {
-        return missionEntityToFlatDto(this);
-    }
-
-    get missionWithFilesDto(): MissionWithFilesDto {
-        return missionEntityToDtoWithFiles(this);
-    }
-
-    get minimumMissionDto(): MinimumMissionDto {
-        return missionEntityToMinimumDto(this);
-    }
 }
-
-export const missionEntityToDto = (mission: Mission): MissionDto => {
-    if (!mission.project) {
-        throw new Error('Mission project is not set');
-    }
-
-    return {
-        ...mission.minimumMissionDto,
-        project: mission.project.projectDto,
-        createdAt: mission.createdAt,
-        tags: mission.tags?.map((tag) => tag.tagDto) || [],
-        updatedAt: mission.updatedAt,
-    };
-};
-
-export const missionEntityToDtoWithCreator = (
-    mission: Mission,
-): MissionWithCreatorDto => {
-    if (!mission.creator) {
-        throw new Error('Mission creator is not set');
-    }
-
-    return {
-        ...missionEntityToDto(mission),
-        creator: mission.creator.userDto,
-    };
-};
-
-export const missionEntityToFlatDto = (mission: Mission): FlatMissionDto => {
-    return {
-        ...missionEntityToDtoWithCreator(mission),
-        filesCount: mission.files?.length || 0,
-        size:
-            mission.files?.reduce(
-                (accumulator, file) => accumulator + (file.size ?? 0),
-                0,
-            ) || 0,
-    };
-};
-
-export const missionEntityToDtoWithFiles = (
-    mission: Mission,
-): MissionWithFilesDto => {
-    if (!mission.files) {
-        throw new Error('Mission files are not set');
-    }
-
-    if (!mission.tags) {
-        throw new Error('Mission creator is not set');
-    }
-
-    return {
-        ...missionEntityToDtoWithCreator(mission),
-        files: mission.files.map((file) => file.fileDto),
-        tags: mission.tags.map((tag) => tag.tagDto),
-    };
-};
-
-export const missionEntityToMinimumDto = (
-    mission: Mission,
-): MinimumMissionDto => {
-    return {
-        name: mission.name,
-        uuid: mission.uuid,
-    };
-};
