+++ common/entities/project/project.entity.ts
@@ -41,55 +41,4 @@ export default class Project extends BaseEntity {
 
     @OneToMany(() => CategoryEntity, (category) => category.project)
     categories?: CategoryEntity[];
-
-    get projectDto(): ProjectDto {
-        return projectEntityToDto(this);
-    }
-
-    get projectWithMissionCountDto(): ProjectWithMissionCountDto {
-        return projectEntityToDtoWithMissionCount(this);
-    }
-
-    get projectWithMissionsDto(): ProjectWithMissionsDto {
-        return projectEntityToDtoWithMissions(this);
-    }
 }
-
-export const projectEntityToDto = (project: Project): ProjectDto => {
-    return {
-        uuid: project.uuid,
-        name: project.name,
-        createdAt: project.createdAt,
-        updatedAt: project.updatedAt,
-        description: project.description,
-    };
-};
-
-export const projectEntityToDtoWithMissionCount = (
-    project: Project,
-): ProjectWithMissionCountDto => {
-    if (project.creator === undefined) {
-        throw new Error('Creator can never be undefined');
-    }
-
-    return {
-        ...projectEntityToDto(project),
-        creator: project.creator.userDto,
-        missionCount: project.missions?.length ?? 0,
-    };
-};
-
-export const projectEntityToDtoWithMissions = (
-    project: Project,
-): ProjectWithMissionsDto => {
-    if (project.creator === undefined) {
-        throw new Error('Creator can never be undefined');
-    }
-
-    return {
-        ...projectEntityToDto(project),
-        creator: project.creator.userDto,
-        requiredTags: project.requiredTags.map((t) => t.tagTypeDto),
-        missions: project.missions?.map((m) => m.flatMissionDto) ?? [],
-    };
-};
