+++ common/entities/tagType/tag-type.entity.ts
@@ -3,7 +3,6 @@ import BaseEntity from '../base-entity.entity';
 import { DataType } from '../../frontend_shared/enum';
 import Project from '../project/project.entity';
 import Tag from '../tag/tag.entity';
-import { TagTypeDto } from '../../api/types/tags/tags.dto';
 
 @Entity()
 export default class TagType extends BaseEntity {
@@ -22,20 +21,5 @@ export default class TagType extends BaseEntity {
 
     @OneToMany(() => Tag, (tag) => tag.tagType)
     tags?: Tag[];
-
-    get tagTypeDto(): TagTypeDto {
-        return tagTypeEntityToDto(this);
-    }
 }
 
-
-export const tagTypeEntityToDto = (tagType: TagType): TagTypeDto => {
-    return {
-        uuid: tagType.uuid,
-        createdAt: tagType.createdAt,
-        updatedAt: tagType.updatedAt,
-        name: tagType.name,
-        description: tagType.description ?? '',
-        datatype: tagType.datatype,
-    };
-}
